refactor: move ArchUnit tests over to `qa/archunit-tests` module
Also removed unused declared dependencies from modules tests moved from
