+++ qa/archunit-tests/src/test/java/io/camunda/client/NamingTest.java
