+++ qa/archunit-tests/src/test/java/io/camunda/zeebe/ForbidWebStereotypeTest.java
