+++ qa/archunit-tests/src/test/java/io/camunda/zeebe/engine/state/migration/MigrationTaskRegistrationTest.java
