+++ qa/archunit-tests/src/test/java/io/camunda/zeebe/gateway/ControllerAuthorizationArchTest.java
