+++ qa/archunit-tests/src/test/java/io/camunda/zeebe/gateway/rest/RestControllerAnnotationTest.java
