+++ qa/archunit-tests/src/test/java/io/camunda/zeebe/protocol/ImmutableProtocolTest.java
